requirements.txt
streamlit
requests
beautifulsoup4
openai>=1.0.0
python-dotenv==0.19.2
